requirements.txt
pandas==2.1.4
numpy==1.26.2
pymongo==4.6.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
scikit-learn==1.3.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
ipykernel==6.28.0
python-dotenv==1.0.0
pytest==7.4.3


